+++ ignite/pkg/chaincmd/chaincmd.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/ignite/cli/ignite/pkg/cmdrunner/step"
 	"github.com/ignite/cli/ignite/pkg/cosmosver"
@@ -34,6 +35,7 @@ const (
 	optionRecover                          = "--recover"
 	optionAddress                          = "--address"
 	optionAmount                           = "--amount"
+	optionFees                             = "--fees"
 	optionValidatorMoniker                 = "--moniker"
 	optionValidatorCommissionRate          = "--commission-rate"
 	optionValidatorCommissionMaxRate       = "--commission-max-rate"
@@ -102,8 +104,8 @@ func (c ChainCmd) Copy(options ...Option) ChainCmd {
 type Option func(*ChainCmd)
 
 func applyOptions(c *ChainCmd, options []Option) {
-	for _, applyOption := range options {
-		applyOption(c)
+	for _, apply := range options {
+		apply(c)
 	}
 }
 
@@ -418,8 +420,8 @@ func (c ChainCmd) GentxCommand(
 	}
 
 	// Apply the options provided by the user
-	for _, applyOption := range options {
-		command = applyOption(command)
+	for _, apply := range options {
+		command = apply(command)
 	}
 
 	command = c.attachChainID(command)
@@ -476,8 +478,21 @@ func (c ChainCmd) ExportCommand() step.Option {
 	return c.daemonCommand(command)
 }
 
+// BankSendOption for the BankSendCommand.
+type BankSendOption func([]string) []string
+
+// BankSendWithFees sets fees to pay along with transaction for the bank send command.
+func BankSendWithFees(fee sdk.Coin) BankSendOption {
+	return func(command []string) []string {
+		if !fee.IsNil() {
+			return append(command, optionFees, fee.String())
+		}
+		return command
+	}
+}
+
 // BankSendCommand returns the command for transferring tokens.
-func (c ChainCmd) BankSendCommand(fromAddress, toAddress, amount string) step.Option {
+func (c ChainCmd) BankSendCommand(fromAddress, toAddress, amount string, options ...BankSendOption) step.Option {
 	command := []string{
 		commandTx,
 	}
@@ -491,6 +506,11 @@ func (c ChainCmd) BankSendCommand(fromAddress, toAddress, amount string) step.Op
 		optionBroadcastMode, flags.BroadcastSync,
 		optionYes,
 	)
+
+	// Apply the options provided by the user
+	for _, apply := range options {
+		command = apply(command)
+	}
 
 	command = c.attachChainID(command)
 	command = c.attachKeyringBackend(command)