+++ changelog.md
@@ -10,6 +10,10 @@
 ### Changes
 
 - [#3529](https://github.com/ignite/cli/pull/3529) Refactor plugin system to use gRPC
+- [#3751](https://github.com/ignite/cli/pull/3751) Rename label to skip changelog check
+- [#3745](https://github.com/ignite/cli/pull/3745) Set tx fee amount as option
+- [#3748](https://github.com/ignite/cli/pull/3748) Change default rpc endpoint to a working one
+- [#3694](https://github.com/ignite/cli/pull/3694) Query and Tx AutoCLI support
 
 ## [`v0.27.1`](https://github.com/ignite/cli/releases/tag/v0.27.1)